It is a project for Issuing Driving License through the concept of Blockchain.
This is a project to Issue License on the basis of your Details Provided. This software also checks if a license has been issued on a identity or not. It also has a portal for the police employees to fine anyone for breaking any traffic law and the fine will be sent to your id.So the person can check it out and pay his fine online.